Move repository specific classes into repository folder and other minor cleanup.
