+++ Eklee.Azure.Functions.GraphQl/InputBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Eklee.Azure.Functions.GraphQl.Repository;
 using GraphQL.Types;
 
 namespace Eklee.Azure.Functions.GraphQl
