+++ Eklee.Azure.Functions.GraphQl/InputBuilderFactory.cs
@@ -1,4 +1,5 @@
-﻿using GraphQL.Types;
+﻿using Eklee.Azure.Functions.GraphQl.Repository;
+using GraphQL.Types;
 
 namespace Eklee.Azure.Functions.GraphQl
 {
