+++ Eklee.Azure.Functions.GraphQl/ModelType.cs
@@ -1,5 +1,4 @@
 ﻿using FastMember;
-using GraphQL.Types;
 using System;
 using System.Collections.Generic;
 using System.Linq;
