+++ Eklee.Azure.Functions.GraphQl/QueryBuilder.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
+using Eklee.Azure.Functions.GraphQl.Repository;
 using GraphQL.Builders;
 using GraphQL.Types;
 using Microsoft.Extensions.Caching.Distributed;
