+++ Eklee.Azure.Functions.GraphQl/Repository/HttpRepository.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Eklee.Azure.Functions.GraphQl
+namespace Eklee.Azure.Functions.GraphQl.Repository
 {
 	public class HttpRepository : IGraphQlRepository
 	{
