+++ Examples/Eklee.Azure.Functions.GraphQl.Example/BusinessLayer/BooksMutation.cs
@@ -1,4 +1,5 @@
 ﻿using Eklee.Azure.Functions.GraphQl.Example.Models;
+using Eklee.Azure.Functions.GraphQl.Repository;
 using GraphQL.Types;
 
 namespace Eklee.Azure.Functions.GraphQl.Example.BusinessLayer
