+++ Examples/Eklee.Azure.Functions.GraphQl.Example/MyModule.cs
@@ -1,6 +1,7 @@
 ﻿using Autofac;
 using Eklee.Azure.Functions.GraphQl.Example.BusinessLayer;
 using Eklee.Azure.Functions.GraphQl.Example.HttpMocks;
+using Eklee.Azure.Functions.GraphQl.Repository;
 using Eklee.Azure.Functions.Http;
 using Microsoft.Extensions.Caching.Distributed;
